refactor: auth
