--- src/middleware/auth-get.ts
@@ -1,6 +1,0 @@
-import { APIGatewayAuthorizerResult, APIGatewayTokenAuthorizerEvent } from "aws-lambda";
-import { authorization } from "./authorization";
-
-export const handler = async (event: APIGatewayTokenAuthorizerEvent): Promise<APIGatewayAuthorizerResult> => {
-    return authorization.valid(event)
-};
